Backend/#11/init registration
* backend #11: Adds user service API with configuration and Docker setup

Implements a user service API that includes database and cache configurations.

Introduces an example environment file, a Makefile for Swagger generation, and updates the Docker Compose setup to include Redis.

Renames the 'username' field to 'nickname' in the user table for clarity and consistency.

Enhances code organization by adding necessary configuration and documentation files to support development and deployment.

* backend #11: fix .gitignore

* backend #11: generated mocks and wrote tests for the registration initialization method

* backend #11: Fixed the way challenges are generated